TeamBoard-backend/apps/boards/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import ProjectViewSet, TaskViewSet, TeamViewSet

router = DefaultRouter()
router.register('teams', TeamViewSet, basename='team')
router.register('projects', ProjectViewSet, basename='project')
router.register('tasks', TaskViewSet, basename='task')

urlpatterns = [
    path('', include(router.urls)),
]