TeamBoard-backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
psycopg2-binary>=2.9
python-dotenv>=1.0
drf-spectacular>=0.27.0
django-cors-headers>=4.0